`@remotion/studio-codemods`: Move Recast composition codemods from `@remotion/studio-server`
