+++ packages/studio-server/src/codemods/duplicate-composition.ts
@@ -1,9 +1,4 @@
-import {duplicateCompositionInSource} from '@remotion/studio-codemods';
-import type {RecastCodemod} from '@remotion/studio-shared';
-import {ensureNamedImport} from '../helpers/imports';
-import {parseAst, serializeAst} from './parse-ast';
-import type {Change} from './recast-mods';
-import {applyCodemod} from './recast-mods';
+export {parseAndApplyCodemod} from '@remotion/studio-codemods';
 
 const getPrettier = async () => {
 	try {
@@ -38,60 +33,4 @@ export const formatOutput = async (tsContent: string) => {
 	});
 
 	return newContents;
-};
-
-export const parseAndApplyCodemod = ({
-	input,
-	codeMod,
-}: {
-	input: string;
-	codeMod: RecastCodemod;
-}): {newContents: string; changesMade: Change[]} => {
-	if (codeMod.type === 'duplicate-composition') {
-		return duplicateCompositionInSource({input, codemod: codeMod});
-	}
-
-	const ast = parseAst(input);
-
-	const {newAst, changesMade} = applyCodemod({
-		file: ast,
-		codeMod,
-	});
-
-	if (changesMade.length === 0) {
-		throw new Error(
-			'Unable to calculate the changes needed for this file. Edit the file manually.',
-		);
-	}
-
-	if (codeMod.type === 'new-composition') {
-		if (codeMod.canvasCapture === null) {
-			ensureNamedImport({
-				ast: newAst,
-				importedName: 'Composition',
-				sourcePath: 'remotion',
-				localName: 'Composition',
-			});
-		}
-
-		ensureNamedImport({
-			ast: newAst,
-			importedName: codeMod.componentName,
-			sourcePath: codeMod.componentImportPath,
-			localName: codeMod.componentName,
-		});
-	}
-
-	if (codeMod.type === 'new-folder') {
-		ensureNamedImport({
-			ast: newAst,
-			importedName: 'Folder',
-			sourcePath: 'remotion',
-			localName: 'Folder',
-		});
-	}
-
-	const output = serializeAst(newAst);
-
-	return {changesMade, newContents: output};
 };
